requirements-example.txt
django==3.2.9
django-environ
djangorestframework
djangorestframework-simplejwt
djoser
django-cors-headers
django-filter
# django-countries
django-autoslug
#django-dbbackup
Pillow
psycopg2-binary
gunicorn
requests
six
typing-extensions
cryptography
flake8
black
isort
wheel
twilio